+++ agentgraph/components/Canvas/Canvas.stories.tsx
@@ -1,0 +1,40 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { Canvas } from './Canvas';
+import { Node } from '../Node'
+
+const meta = {
+  title: 'AgentGraph/Canvas',
+  component: Canvas,
+  parameters: { layout: 'fullscreen' },
+  tags: ['autodocs'],
+  argTypes: {},
+} satisfies Meta<typeof Canvas>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Empty: Story = {}
+
+// Put a node in the canvas
+export const WithNode: Story = {
+  args: {
+    children: (
+      <Node step={{
+        agent_id: 1,
+        category: 'validation',
+        created_at: '2021-08-31T15:00:00.000Z',
+        description: "Ensure input is a valid chat message",
+        entry_type: 'input',
+        error_message: "Could not validate input",
+        id: 1,
+        name: "Validate Input",
+        order: 1,
+        success_action: "next_node",
+        task_id: 1,
+        updated_at: '2021-08-31T15:00:00.000Z',
+      }}
+      />
+    )
+  }
+}
